[FIX] google_calendar: prevent to sync recurrencies with false rrule.
Before this commit, when a recurrency was not properly populated, it could be synced.
Google could accomodate with these records but when they are synched again later after a cleaup or in another Odoo instance , they are created with default value.
The typical error encoutered would be:
"You have to choose at least one day in the week"

Because the recurrency would be created with a default frequency (weekly) and no default day selected.
This commit prevent to sync such records.

closes odoo/odoo#69619

Opw: 2450259 (forward port + adaptation/fixes to 14.2 changes)
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>